third_party/blink/web_tests/external/wpt/fetch/private-network-access/shared-worker.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public to local: failure.
  assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
[FAIL] treat-as-public to private: failure.
  assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
Harness: the test ran to completion.
